+++ packages/forest_admin_datasource_mambu_payments/spec/forest_admin_datasource_mambu_payments/plugins/smart_actions/trigger_payee_verification_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe ForestAdminDatasourceMambuPayments::Plugins::TriggerPayeeVerification do
+RSpec.describe ForestAdminDatasourceMambuPayments::Plugins::SmartActions::TriggerPayeeVerification do
   let(:client) { instance_double(ForestAdminDatasourceMambuPayments::Client) }
   let(:datasource) { instance_double(ForestAdminDatasourceMambuPayments::Datasource, client: client) }
   let(:result_builder) { ForestAdminDatasourceCustomizer::Decorators::Action::ResultBuilder.new }
