+++ packages/forest_admin_datasource_mambu_payments/spec/forest_admin_datasource_mambu_payments/plugins/smart_actions/update_external_account_spec.rb
@@ -1,9 +1,9 @@
-RSpec.describe ForestAdminDatasourceMambuPayments::Plugins::UpdateInternalAccount do
+RSpec.describe ForestAdminDatasourceMambuPayments::Plugins::SmartActions::UpdateExternalAccount do
   let(:client) { instance_double(ForestAdminDatasourceMambuPayments::Client) }
   let(:datasource) { instance_double(ForestAdminDatasourceMambuPayments::Datasource, client: client) }
   let(:result_builder) { ForestAdminDatasourceCustomizer::Decorators::Action::ResultBuilder.new }
   let(:collection_customizer) { ForestAdminDatasourceMambuPayments::PluginSupport::FakeCollection.new }
-  let(:record_id_field) { 'mambu_internal_account_id' }
+  let(:record_id_field) { 'mambu_external_account_id' }
 
   def register(opts = {})
     described_class.new.run(nil, collection_customizer,
@@ -11,16 +11,21 @@ RSpec.describe ForestAdminDatasourceMambuPayments::Plugins::UpdateInternalAccoun
     collection_customizer.registered[opts[:action_name] || described_class::NAME]
   end
 
-  it 'PATCHes /internal_accounts/{id} with only the filled fields' do
+  it 'PATCHes /external_accounts/{id} with only the filled fields' do
     action = register
-    allow(client).to receive(:update_internal_account)
+    allow(client).to receive(:update_external_account)
     context = ForestAdminDatasourceMambuPayments::PluginSupport::FakeContext.new(
-      records: [{ record_id_field => 'ia_1' }],
-      form_values: { 'Name' => 'New', 'Holder name' => '', 'Account number' => '' }
+      records: [{ record_id_field => 'ea_1' }],
+      form_values: { 'Holder name' => 'New', 'Account number' => '', 'Bank code' => '' }
     )
 
     action.execute.call(context, result_builder)
 
-    expect(client).to have_received(:update_internal_account).with('ia_1', 'name' => 'New')
+    expect(client).to have_received(:update_external_account).with('ea_1', 'holder_name' => 'New')
+  end
+
+  it 'raises ArgumentError without :record_id_field' do
+    expect { described_class.new.run(nil, collection_customizer, datasource: datasource) }
+      .to raise_error(ArgumentError, /record_id_field/)
   end
 end
